+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0 - 2023-10-21
+### Changed
+- settings are now admin settings
+- settings can be delegated with admin privileges
+- all settings are stored encrypted in appconfig
+
 ## 0.1.0 - 2022-12-30
 - initial Nextcloud app store release